+++ CHANGELOG.md
@@ -8,6 +8,8 @@
   - Handle data-URI-style inline PHOTO as used in vCard4
   - Use VCard conversion to handle v4 properties such as KIND=group for which extensions are used in v3 vCards
 - Fix: Do not attempt to download photos from URIs with unsupported URI scheme (supported are http and https) (#411)
+- When an invalid value is submitted in one of the settings forms, RCMCardDAV will now abort with an error. Previous
+  versions of RCMCardDAV would ignore the value when possible an instead keep the existing value or use a default value.
 
 ## Version 5.0.0-beta1 (to 4.4.4)