+++ CHANGELOG.md
@@ -10,6 +10,7 @@
 - Fix: Do not attempt to download photos from URIs with unsupported URI scheme (supported are http and https) (#411)
 - When an invalid value is submitted in one of the settings forms, RCMCardDAV will now abort with an error. Previous
   versions of RCMCardDAV would ignore the value when possible an instead keep the existing value or use a default value.
+- Backend will refuse Frontend request to remove user account created from preset
 
 ## Version 5.0.0-beta1 (to 4.4.4)